requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.1; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
anyio==3.7.1; python_version >= '3.7'
asgiref==3.7.2; python_version >= '3.7'
attrs==23.1.0; python_version >= '3.7'
azure-common==1.1.28
azure-core==1.29.6; python_version >= '3.7'
azure-identity==1.15.0
azure-search-documents==11.4.0b8
backoff==2.2.1; python_version >= '3.7'
bcrypt==4.1.2; python_version >= '3.7'
beautifulsoup4==4.12.2
cachetools==5.3.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cryptography==41.0.7; python_version >= '3.7'
curl-cffi==0.5.10; python_version >= '3.7'
dataclasses-json==0.6.3; python_version >= '3.7' and python_version < '4.0'
deprecated==1.2.14; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
distro==1.8.0; python_version >= '3.6'
docarray==0.39.1
duckduckgo-search==4.1.0
fastapi==0.105.0; python_version >= '3.8'
filelock==3.13.1; python_version >= '3.8'
flatbuffers==23.5.26
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2023.12.2; python_version >= '3.8'
google-auth==2.25.2; python_version >= '3.7'
googleapis-common-protos==1.62.0; python_version >= '3.7'
greenlet==3.0.2; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
grpcio==1.60.0; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.2; python_version >= '3.8'
httptools==0.6.1
httpx==0.25.2; python_version >= '3.8'
huggingface-hub==0.19.4; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.6; python_version >= '3.5'
importlib-metadata==6.11.0; python_version >= '3.8'
importlib-resources==6.1.1; python_version >= '3.8'
isodate==0.6.1
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
kubernetes==28.1.0; python_version >= '3.6'
langchain==0.0.351
langchain-community==0.0.4; python_version < '4.0' and python_full_version >= '3.8.1'
langchain-core==0.1.1; python_version < '4.0' and python_full_version >= '3.8.1'
langserve==0.0.36
langsmith==0.0.72
lxml==4.9.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markdown-it-py==3.0.0; python_version >= '3.8'
marshmallow==3.20.1; python_version >= '3.8'
mdurl==0.1.2; python_version >= '3.7'
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
msal==1.26.0; python_version >= '2.7'
msal-extensions==1.1.0; python_version >= '3.7'
multidict==6.0.4; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
numpy==1.26.2; python_version >= '3.9'
oauthlib==3.2.2; python_version >= '3.6'
onnxruntime==1.16.3
openai==1.5.0
opentelemetry-api==1.21.0; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-common==1.21.0; python_version >= '3.7'
opentelemetry-exporter-otlp-proto-grpc==1.21.0; python_version >= '3.7'
opentelemetry-instrumentation==0.42b0; python_version >= '3.7'
opentelemetry-instrumentation-asgi==0.42b0; python_version >= '3.7'
opentelemetry-instrumentation-fastapi==0.42b0; python_version >= '3.7'
opentelemetry-proto==1.21.0; python_version >= '3.7'
opentelemetry-sdk==1.21.0; python_version >= '3.7'
opentelemetry-semantic-conventions==0.42b0; python_version >= '3.7'
opentelemetry-util-http==0.42b0; python_version >= '3.7'
orjson==3.9.10; python_version >= '3.8'
overrides==7.4.0; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
portalocker==2.8.2; platform_system == 'Windows'
posthog==3.1.0
protobuf==4.25.1; python_version >= '3.8'
pulsar-client==3.3.0
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic==1.10.13
pygments==2.17.2; python_version >= '3.7'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
pypdf==3.17.3
pypika==0.48.9
pyreadline3==3.4.1; python_version >= '3.8' and sys_platform == 'win32'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.0
pywin32==306; platform_system == 'Windows'
pyyaml==6.0.1; python_version >= '3.6'
requests==2.31.0
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==13.7.0; python_full_version >= '3.7.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
setuptools==69.0.2; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlalchemy==2.0.23; python_version >= '3.7'
sse-starlette==1.8.2
starlette==0.27.0; python_version >= '3.7'
sympy==1.12; python_version >= '3.8'
tenacity==8.2.3; python_version >= '3.7'
tokenizers==0.15.0; python_version >= '3.7'
tqdm==4.66.1; python_version >= '3.7'
typer==0.9.0; python_version >= '3.6'
types-requests==2.31.0.6; python_version >= '3.7'
types-urllib3==1.26.25.14
typing-extensions==4.9.0; python_version >= '3.8'
typing-inspect==0.9.0
urllib3==1.26.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
uvicorn[standard]==0.24.0.post1; python_version >= '3.8'
watchfiles==0.21.0
websocket-client==1.7.0; python_version >= '3.8'
websockets==12.0
wrapt==1.16.0; python_version >= '3.6'
yarl==1.9.4; python_version >= '3.7'
zipp==3.17.0; python_version >= '3.8'
